feat(dm): surface combat turn header
