+++ client/src/App.scss
@@ -430,6 +430,24 @@
   line-height: 1.2;
 }
 
+.zombies-dm-page__combat-header {
+  position: absolute;
+  top: calc(env(safe-area-inset-top, 0px) + clamp(6rem, 12vh, 8.5rem));
+  left: 50%;
+  transform: translateX(-50%);
+  width: min(92vw, 1100px);
+  padding: 0 clamp(0.75rem, 2vw, 1.5rem);
+  pointer-events: auto;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 0.75rem;
+}
+
+.zombies-dm-page__combat-header .combat-turn-header__active-indicator {
+  max-width: 100%;
+}
+
 .zombies-dm-bottom-bar {
   position: fixed;
   inset: auto 0 0 0;
